petri/pyproject.toml
[project]
name = "test-petri"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "petri",
    "python-dotenv>=1.2.1",
]

[tool.uv.sources]
petri = { git = "https://github.com/safety-research/petri" }
